+++ CHANGELOG.md
@@ -19,6 +19,11 @@
           type: string
   ```
 
+## Bug Fixes
+
+- When using `enum` in conjunction with `x-nullable` in a schema, this will now
+  result in the `null` value being present in schema if it isn't already.
+
 # 0.19.0
 
 ## Enhancements